+++ packages/frontend-framework/src/components/navbar/tabs/simpleNavbarTab/simpleNavbarTabUnfolded.tsx
@@ -20,6 +20,7 @@ import React from "react";
 import {SvgIcon} from "../svgIcon";
 import {NestedNavbarTabProps} from "./simpleNavbarTab";
 import "../tabs.css";
+import {DefaultIcon} from "../defaultIcon";
 
 export const SimpleNavbarTabUnfolded = (props: NestedNavbarTabProps) => {
   let className = "default-nav-element-unfolded default-tab-unfolded flex ";
@@ -36,7 +37,11 @@ export const SimpleNavbarTabUnfolded = (props: NestedNavbarTabProps) => {
       onMouseLeave={() => props.setHovering(false)}
     >
       <div className="flex align-items-center">
-        <SvgIcon color={props.iconColor} element={props.icon} />
+        {props.icon ? (
+          <SvgIcon color={props.iconColor} element={props.icon} />
+        ) : (
+          <DefaultIcon color={props.iconColor} />
+        )}
         <span id="navbar-tab-name">{props.name}</span>
       </div>
     </div>