+++ packages/frontend-framework/CHANGELOG.md
@@ -1,3 +1,23 @@
+## [@iavofficial/frontend-framework-v1.4.0-dev.1](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-v1.3.1-dev.2...@iavofficial/frontend-framework-v1.4.0-dev.1) (2025-02-20)
+
+### Features
+
+* add class for div ([83e6273](https://github.com/iavofficial/IAVFrontendFramework/commit/83e6273aceb680ae7f8241a081a4f1bccf467992))
+* add user icon as ReactElement ([184508f](https://github.com/iavofficial/IAVFrontendFramework/commit/184508fb8ab6fd3582de6cae097ec70d52fcf54b))
+* add width and height for icon ([c680e0b](https://github.com/iavofficial/IAVFrontendFramework/commit/c680e0bf1c10b3e23485cca08a18932a1f594c91))
+
+## [@iavofficial/frontend-framework-v1.3.1-dev.2](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-v1.3.1-dev.1...@iavofficial/frontend-framework-v1.3.1-dev.2) (2025-02-20)
+
+### Bug Fixes
+
+* add background in mainVíew.tsx ([eabf8d3](https://github.com/iavofficial/IAVFrontendFramework/commit/eabf8d3341ef2ce75c834c8a8a1bc4d0c72c9473))
+
+## [@iavofficial/frontend-framework-v1.3.1-dev.1](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-v1.3.0...@iavofficial/frontend-framework-v1.3.1-dev.1) (2025-02-20)
+
+### Bug Fixes
+
+* add default icon ([6e3fce3](https://github.com/iavofficial/IAVFrontendFramework/commit/6e3fce37ff99e3bd9f27e8412022767c2e98a60d))
+
 ## [1.3.0](https://github.com/iavofficial/IAVFrontendFramework/compare/v1.2.0...v1.3.0) (2025-02-10)
 
 ### Features